Rename nessie-quarkus-cli to nessie-server-admin-tool
Also move code under ./tools/server-admin

This is a follow-up to #8245
Closes #8481
