+++ .github/workflows/newer-java.yml
@@ -59,7 +59,7 @@ jobs:
           spotlessCheck \
           compileAll \
           -x :nessie-quarkus:compileAll \
-          -x :nessie-quarkus-cli:compileAll \
+          -x :nessie-server-admin-tool:compileAll \
           -x :nessie-events-quarkus:compileAll \
           --scan
 
@@ -67,7 +67,7 @@ jobs:
       run: |
         ./gradlew \
           :nessie-quarkus:compileAll \
-          :nessie-quarkus-cli:compileAll \
+          :nessie-server-admin-tool:compileAll \
           :nessie-events-quarkus:compileAll \
           --scan
