+++ gradle/projects.main.properties
@@ -30,7 +30,6 @@ nessie-keycloak-testcontainer=testing/keycloak-container
 nessie-nessie-testcontainer=testing/nessie-container
 nessie-object-storage-mock=testing/object-storage-mock
 nessie-quarkus-auth=servers/quarkus-auth
-nessie-quarkus-cli=servers/quarkus-cli
 nessie-quarkus-common=servers/quarkus-common
 nessie-quarkus=servers/quarkus-server
 nessie-quarkus-tests=servers/quarkus-tests
@@ -43,6 +42,7 @@ nessie-services-config=servers/services-config
 nessie-services-bench=servers/services-bench
 nessie-server-store=servers/store
 nessie-server-store-proto=servers/store-proto
+nessie-server-admin-tool=tools/server-admin
 nessie-content-generator=tools/content-generator
 nessie-protobuf-relocated=tools/protobuf-relocated
 nessie-storage-uri=tools/storage/uri
