+++ servers/quarkus-server/src/main/resources/application.properties
@@ -67,7 +67,7 @@ nessie.server.send-stacktrace-to-client=false
 #
 # Legacy version store types: INMEMORY, ROCKS, DYNAMO, MONGO, TRANSACTIONAL. If you are using
 # one of these legacy version store types, migrate your existing repositories to the new version
-# store types using the nessie-quarkus-cli tool's export/import functionality.
+# store types using the nessie-server-admin-tool's export/import functionality.
 nessie.version.store.type=IN_MEMORY
 
 # Object cache size as a value relative to the JVM's max heap size. The `cache-capacity-fraction-adjust-mb`
