+++ tools/server-admin/src/intTest/java/org/projectnessie/tools/admin/cli/QuarkusCliTestProfilePersistMongo.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2024 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.tools.admin.cli;
+
+import java.util.Collections;
+import java.util.List;
+import org.projectnessie.quarkus.tests.profiles.QuarkusTestProfilePersistMongo;
+
+public class QuarkusCliTestProfilePersistMongo extends QuarkusTestProfilePersistMongo {
+  @Override
+  public List<TestResourceEntry> testResources() {
+    // In CLI tests external resources (MongoDB) are managed by NessieCliTestExtension
+    return Collections.emptyList();
+  }
+}
