+++ tools/server-admin/src/main/resources/application.properties
