+++ tools/server-admin/src/main/resources/org/projectnessie/tools/admin/cli/version.properties
